backend/requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bravado==11.0.3
bravado-core==6.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.8
Django==4.2.9
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
filelock==3.13.1
fqdn==1.5.1
idna==3.6
isoduration==20.11.0
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
monotonic==1.6
msgpack==1.0.7
numpy==1.26.3
pandas==2.2.0
platformdirs==4.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyupbit==0.2.33
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
setuptools==69.0.3
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
swagger-spec-validator==3.0.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
upbit-client==1.3.4.1
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.25.0
webcolors==1.13
websockets==12.0
